+++ src/Client/ClientCore/Entities/TemplateEntityDataLoader.cs
@@ -1,0 +1,51 @@
+// Sovereign Engine
+// Copyright (c) 2024 opticfluorine
+// 
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+// 
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+// 
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+using Castle.Core.Logging;
+using Sovereign.EngineCore.Entities;
+
+namespace Sovereign.ClientCore.Entities;
+
+/// <summary>
+///     Responsible for loading template entity data in the client.
+/// </summary>
+public class TemplateEntityDataLoader
+{
+    private readonly EntityDefinitionProcessor definitionProcessor;
+    private readonly IEntityFactory factory;
+
+    public TemplateEntityDataLoader(IEntityFactory factory, EntityDefinitionProcessor definitionProcessor)
+    {
+        this.factory = factory;
+        this.definitionProcessor = definitionProcessor;
+    }
+
+    public ILogger Logger { private get; set; } = NullLogger.Instance;
+
+    /// <summary>
+    ///     Loads a set of template entity data into the entity and component tables.
+    /// </summary>
+    /// <param name="data">Data.</param>
+    public void Load(TemplateEntityData data)
+    {
+        foreach (var def in data.TemplateEntityDefinitions)
+        {
+            definitionProcessor.ProcessDefinition(def);
+        }
+
+        Logger.InfoFormat("Loaded {0} template entities from server.", data.TemplateEntityDefinitions.Count);
+    }
+}